# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Web Media Hub - Multi-store clothing management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products (newest first)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [Top, Bottom, "Full Outfit", Functional]
        - in: query
          name: storeId
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /products/{id}/like:
    post:
      operationId: likeProduct
      tags: [products]
      summary: Toggle like on a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /products/upload-image:
    post:
      operationId: uploadProductImage
      tags: [products]
      summary: Upload product image (base64)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUploadInput"
      responses:
        "200":
          description: Uploaded image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResponse"
...
  /store:
    get:
      operationId: getStore
      tags: [store]
      summary: Get store configuration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Store info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [store]
      summary: Create/save store configuration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "200":
          description: Store saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
    patch:
      operationId: updateStore
      tags: [store]
      summary: Update store configuration (after save/lock)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Store updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    Product:
      type: object
      required: [id, name, discountPrice, actualPrice, productType, sizes, images, likeCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        discountPrice:
          type: number
        actualPrice:
          type: number
        functionCategory:
          type: ["string", "null"]
        productType:
          type: string
          enum: [Top, Bottom, "Full Outfit", Functional]
        sizes:
          type: array
          items:
            type: string
        likeCount:
          type: integer
        storeId:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, discountPrice, actualPrice, productType, sizes, images]
      properties:
        name:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        discountPrice:
          type: number
        actualPrice:
          type: number
        functionCategory:
          type: string
        productType:
          type: string
          enum: [Top, Bottom, "Full Outfit", Functional]
        sizes:
          type: array
          items:
            type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        discountPrice:
          type: number
        actualPrice:
          type: number
        functionCategory:
          type: string
        productType:
          type: string
          enum: [Top, Bottom, "Full Outfit", Functional]
        sizes:
          type: array
          items:
            type: string
    LikeResponse:
      type: object
      required: [likeCount]
      properties:
        likeCount:
          type: integer
    ImageUploadInput:
      type: object
      required: [imageData, fileName]
      properties:
        imageData:
          type: string
          description: Base64 encoded image data
        fileName:
          type: string
    ImageUploadResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
...
    Store:
      type: object
      required: [id, name, isLocked]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        openingTime:
          type: ["string", "null"]
        openDays:
          type: ["string", "null"]
        bannerImage:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        publicSlug:
          type: ["string", "null"]
        isLocked:
          type: boolean
        createdAt:
          type: string
    StoreInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
        whatsappNumber:
          type: string
        openingTime:
          type: string
        openDays:
          type: string
        bannerImage:
          type: string
        description:
          type: string
    StoreUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        whatsappNumber:
          type: string
        openingTime:
          type: string
        openDays:
          type: string
        bannerImage:
          type: string
        description:
          type: string
...