# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Web Media Hub - Multi-store clothing management API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputProductType } from './productInputProductType';

export interface ProductInput {
  name: string;
  description?: string;
  images: string[];
  discountPrice: number;
  actualPrice: number;
  functionCategory?: string;
  productType: ProductInputProductType;
  sizes: string[];
}